utils/log: Do all log formatting in src/utils/loggers.ts
This commit is a work-around for targets that only display the first
arguments of console.log-like calls. Here a basic transformation of
object to a string is also done, with a size limit to not overload logs
with very long objects stringified into JSON.

I don't know if this will still be needed in the future but it's a nice
thing to have.