eme: re-implement getMediaKeySystemAccess through the usual operators
The previous implementation was done by defining an Observable from
scratch and subscribing to an observable version of
`requestMediaKeySystemAcess` inside.
Also, some state had to be kept to know if the Observable had been
unsubscribed or not inside.

This wasn't readable, even more considering that we could simply
implement it with the usual operator and let RxJS handle the teardown
itself.