dash/parsers: implement SegmentTemplate inheritance when parsing the MPD
Fixes #767

This PR implements a new DASH behavior when parsing a MPD:

SegmentTemplate at different levels (for example at the `Period`,
`AdaptationSet` _AND_ `Representation` levels) are all merged into one
once converted into our internal `Manifest` structure.

This follows multiple MPDs we have seen recently that present this
particularity. This should not break any other use-case.

The way I implemented it is by making the node_parsers part of our DASH
parser (which has the task to take the MPD from its original `document`
form and transform it to a JS object, very close to the original but
with attributes in the right type - called the "intermediate
representation") even dumber than before by removing the previous
explicit distinction between a "SegmentTemplate" and a "SegmentTimeline".

Now, a timeline-based index will just be a SegmentTemplate one
containing a `timelineParser` attribute (we are thus here much closer to
the original logic of a MPD where a SegmentTimeline is just a child of a
SegmentTemplate).

In the node_parsers, I also removed the default values calculated for
some attributes of the SegmentTemplate/SegmentTimeline/SegmentBase
code. This is because it entered in conflict with the merging happening
much later in the parser code. When that merge happen, we wouldn't be
able to make the difference between a default attribute and an explicit
one, despite the fact that default one can be overwritten when
inheriting parent SegmentTemplates intermediate representation.

As the merging code complexify a little the final parsing of a
representation (in parse_representations.ts), I created a new file -
parse_representation_index.ts - whose tasks is just to parse a
RepresentationIndex. Its exported function has the same signature than
the ones in the other parse_XXX.ts files: the corresponding intermediate
representation as first argument and supplementary context as a second
argument.

I also added some integration tests to check the feature.
I added two test contents for the two cases encountered in the real
world:
  1. A MPD with both a SegmentTemplate in a Period indicating the URL
     forms and one in the AdaptationSet indicating the SegmentTimeline +
     timescale.
  2. A MPD with a SegmentTemplate in an AdaptationSet indicating the
     timescale and another in the Representation indicating the
     SegmentTimeline and urls.