fetchers: the ObservablePrioritizer's _interrupt pending task now take a task in argument
Its previous argument was the index in a _pendingTasks array. After some
discussion, we found out that it may be dangerous in the future to pass
an index around.

It is safer to give the full task: even if the _pendingTasks array is
mutated in the mean time, the task can still be retrieved.

The downside is that we will perform more computation than needed. But
we're only talking about a very very minor performance penalty so it has
been decided that the safer approach wins.