buffers: create new RepresentationBuffer as soon as possible
The previous logic waited until a previous `RepresentationBuffer`'s
Observable completed before creating a new one when switching quality.

This is fine but it could mean some milliseconds lost without a request
(for example when the old RepresentationBuffer is waiting on segments it
has pushed to be completely pushed.).

I updated this behavior so that now a new `RepresentationBuffer` is
created as soon as the `"buffer-terminating" is received from the
previous `RepresentationBuffer`. This should not bring much conflicts as
once it sent a `"buffer-terminating"` event, a `RepresentationBuffer`
just waits for segments to be fully pushed (it will still send the
corresponding events when done).