buffers: do not load again segments that are being pushed to the SourceBuffer
Following some updates and optimizations, it is now possible that
multiple `RepresentationBuffer` for the same `Period` are alive at the
same time.

A side-effect to that new behavior is that it is now possible that a
given `RepresentationBuffer` is still pushing a segment which begins
and ends at the same time than another segment another
`RepresentationBuffer` is considering to download.

Here, we however might still want to load again that segment from the
new RepresentationBuffer, for example to speed-up quality change (what
we call fast-switching) in the case that new `RepresentationBuffer` is
linked to a `Representation` with a higher-bitrate

So this commit both checks for each segment a `RepresentationBuffer`
wants to download if:
  1. there's already a segment that is being pushed at roughly the
     same time
  2. if at least one of those segments does not need to be replaced.

If both are true, we do not load the corresponding segment again.